Expose the initializeApp function from Firebase
Without this, using Firebase from the application consuming this library does
not work and results in an error about not having an app initialized.

This also makes sense so that this library fully encapsulates the firebase
library and the consuming app doesn't even need to import anything directly from
the firebase module.