Add a simple example app with create-react-app
This demonstrates how to use this library with a simple example.
The only small thing is the lack of this library in the package.json file.
I'm not sure what is best for that because in development I used `yarn link`.

Reminder, to use `yarn link`:
- In the library code, run `yarn link`
- In the consuming app code, run `yarn link fire-event-store-react`

This will symlink the library into the consuming apps `node_modules` directory